fix: restart

+++ scripts/start_gl_nohup.sh
@@ -2,10 +2,15 @@ STDIR=`dirname $0`
 
 cd $STDIR/../
 
+PRUNE=${1:-"YES"}
+
+
 # take down OMT
 docker-compose down
-docker system prune -a -f
 mv out.txt out.old
+if [ "$PRUNE" == "YES" ]; then
+  docker system prune -a -f
+fi
 
 # restart OMT
 cmd="docker-compose up -d tileserver-gl >> out.txt 2>&1"
